refactor(ui): fix typo in IUser.ts name
